leds: Add the "input" trigger, for pwr_led
commit 875a7e394f836dedfd7fdff1e15ae1be51d54e47 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The "input" trigger makes the associated GPIO an input.  This is to support
the Raspberry Pi PWR LED, which is driven by external hardware in normal use.

N.B. pwr_led is not available on Model A or B boards.

leds-gpio: Implement the brightness_get method

The power LED uses some clever logic that means it is driven
by a voltage measuring circuit when configured as input, otherwise
it is driven by the GPIO output value. This patch wires up the
brightness_get method for leds-gpio so that user-space can monitor
the LED value via /sys/class/gpio/led1/brightness. Using the input
trigger this returns an indication of the system power health,
otherwise it is just whatever value the trigger has written most
recently.

See: https://github.com/raspberrypi/linux/issues/1064
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>