video: bcm2708_fb: Disable FB if no displays found
commit 2dee5238a25d0d7f27cbccd68b7e1d3fa16edcc8 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

If the firmware hasn't detected a display, the driver would assume
one display was available, but because it had failed to retrieve the
display size it would try to allocate a zero-sized buffer.

Avoid the allocation failure by bailing out early if no display is
found.

See: https://github.com/raspberrypi/linux/issues/3598

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>