media: bcm2835-unicam: change minimum number of vb2_queue buffers to 1
commit 435a06c50538b76333b15242379e35f802322c89 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Since the unicam driver was modified to write to a dummy buffer when no
user-supplied buffer is available, it can now write to and return a
buffer even when there's only a single one. Enable this by changing the
min_buffers_needed in the vb2_queue; it will be useful for enabling
still captures without allocating more memory than absolutely necessary.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>