firmware: raspberrypi: Add support for tryonce reboot flag
commit 777a6a08bcf8f5f0a0086358dc66d8918a0e1c57 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Define a new mailbox (SET_REBOOT_FLAGS) which may be used to
pass optional flags to the Raspberry Pi firmware that changes
the behaviour of the bootloader and firmware during a reboot.

Currently this just defines the 'tryboot' flag which causes
the firmware to load tryboot.txt instead config.txt. This
alternate configuration file can be used to specify the
path of an alternate firmware and kernels allowing a fallback
mechanism to be implemented for OS upgrades.

Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>