media: bcm2835-unicam: Return early from stop_streaming() if stopped
commit 0db8160cf18f92bd666d018bb0a1fed4c8b63769 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

clk_disable_unprepare() is called unconditionally in stop_streaming().
This is incorrect in the cases where start_streaming() fails, and
unprepares all clocks as part of the failure cleanup. To avoid this,
ensure that clk_disable_unprepare() is only called in stop_streaming()
if the clocks are in a prepared state.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>