kbuild: Silence unavoidable dtc overlay warnings
commit c3b75d7bb31ecd075ff2d5ebd7425e3474090cb5 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Much effort has been put into finding ways to avoid warnings from dtc
about overlays, usually to do with the presence of #address-cells and
size-cells, but not exclusively so. Since the issues being warned about
are harmless, suppress the warnings to declutter the build output and
to avoid alarming users.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>