drm/vc4: hvs: Force modeset on gamma lut change
commit 50a879cfdb87baad4edb50f7b443177a592998ed from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The HVS Gamma block can only be updated when idle, so we need to disable
the HVS channel when the gamma property is set in an atomic commit.

Since the pixelvalve cannot have its assigned channel halted without
stalling unless it's disabled as well, in our case that means forcing a
full disable / enable cycle on the pipeline.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>