power: rpi-poe: Add option of being created by MFD or FW
commit 50aa893f0862b33ae14367454b0fa10407af2b7c from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The firmware can only use I2C0 if the kernel isn't, therefore
with libcamera and DRM using it the PoE HAT fan control needs
to move to the kernel.

Add the option for the driver to be created by the PoE HAT core
MFD driver, and use the I2C regmap that provides to control fan
functions.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>