drivers: staging: bcm2835-isp: Do not cleanup mmal vcsm buffer on stop_streaming
commit 4abb2b331efce21c0e9ba8b9968f1fe1f8b4acd7 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

On stop_streaming() the vcsm buffer handle gets released by the buffer cleanup
code.  This will subsequently cause and error if userland re-queues the same
buffer on the next start_streaming() call.

Remove this cleanup code and rely on the vb2_ops->buf_cleanup() call to do the
cleanups instead.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>