drm/atomic: If margins are updated, update all planes.
commit f6d8271436e2589629ed6f3a8a85c3bde53353d6 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Margins may be implemented by scaling the planes, but as there
is no way of intercepting the set_property for a standard property,
and all planes are checked in drm_atomic_check_only before the
connectors, there's now way to add the planes into the state
from the driver.

If the margin properties change, add all corresponding planes to
the state.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>