media: i2c: imx258: Change register settings for variants of the sensor
commit af972efc5fbaa4bd7bb26ffd868c54b71d49b4cc from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Sony have advised that there are variants of the IMX258 sensor which
require slightly different register configuration to the mainline
imx258 driver defaults.

There is no available run-time detection for the variant, so add
configuration via the DT compatible string.

The Vision Components imx258 module supports PDAF, so add the
register differences for that variant

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>