drm/vc4: hdmi: Add property to allow manual config of RGB or YCbCr
commit 5ba39f5b08d6df4aadbe60768b82b5e2d405c0d5 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Add a custom property "Output format" that allows the overriding
of the default colourspace choice in the way that the old
firmware hdmi_pixel_encoding property did. If the chosen format is not
supported, then it will still drop back to the older behaviour.

This won't be acceptable to upstream, but it adds back the missing
functionality of hdmi_pixel_encoding.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>