media: i2c: arducam_64mp: Modify the line length of 1280x720 resolution
commit 506999220463785d41d971346cc4619e09da708d from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Arducam 64MP has specific requirements for the line length, and if these
conditions are not met, the camera will not function properly. Under the
previous configuration, once a stream off operation is performed, the
camera will not output any data, even if a stream on operation is
performed. This prevents us from switching from 1280x720 to another
resolution.

Signed-off-by: Lee Jackson <lee.jackson@arducam.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>