drivers: media: bcm2835_unicam: Improve frame sequence count handling
commit ea9953fe8cff5e0b8e456f36558b054ade871ffe from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Ensure that the frame sequence counter is incremented only if a previous
frame start interrupt has occurred, or a frame start + frame end has
occurred simultaneously.

This corresponds the sequence number with the actual number of frames
produced by the sensor, not the number of frame buffers dequeued back
to userland.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>