media: platform: microchip: Fix OSD boundary exceeding the frame boundary
commit 70cc4843ae4d114aa2377c0bf39c688b0c34515d from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Using the v4l2 command the OSD feature can be disabled, but actually it is
not disabled in the hardware, instead it is shifted to out of resolution
boundary(frame boundary), so that it won't appear on the screen. But when
the resoution is increased, the OSD reappears on the screen even though
it is disabled.

To fix this a separate register was added to the IP, so that the OSD will
get disabled in the hardware itself.

Signed-off-by: Vattipalli Praveen <praveen.kumar@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>