iommu: Don't reserve IOVA when address and size are zero
commit 3011093b02e3f7a2526dfa632ec20fd4ce99e793 from
https://nv-tegra.nvidia.com/r/3rdparty/canonical/linux-jammy.git jetson_36.3

BugLink: https://bugs.launchpad.net/bugs/2034622

When the bootloader doesn't setup the framebuffers, their address and
size are zero in iommu-addresses property. If we intend to use display
driver in kernel without framebuffer then it's causing display mappings
to fail as IOVA is reserved with size and address as zero.
Fix this by checking for size and address of IOVA and skipping the
reservation if both are 0.

http://nvbugs/4222942

Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>