media/rpivid: Allow use of iommu in rpivid
commit e2dd342f4ab1ff563c4ff9171252a1dea947669d from
https://github.com/raspberrypi/linux.git rpi-6.6.y

In order to use iommu on hevc set dma mask_and_coherent in probe.
I am assured dma_set_mask_and_coherent is benign on Pi4 (which has
no iommu) and it seems to be so in practice.
Also adds a bit of debug to make internal buffer allocation failure
easier to spot in future

Signed-off-by: John Cox <jc@kynesim.co.uk>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>