dt-bindings: interrupt-controller: document PolarFire SoC's gpio interrupt mux
commit 8fc0001091cf829fa57f3ec8a8ada311a5bbf6a0 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

On PolarFire SoC there are more GPIO interrupts than there are interrupt
lines available on the PLIC, and a runtime configurable mux is used to
decide which interrupts are assigned direct connections to the PLIC &
which are relegated to sharing a line.
This mux is, in our reference configuration, written by platform
firmware during boot based on the FPGA's configuration.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>