drivers: media: imx477: Set horizontal binning when disabling the scaler
commit 0113f3172c3f7119c3f9aa50614c7e53412c54e7 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The horizontal scaler has been disabled but actually the sensor is not
binning horizontally, resulting in images that are stretched 2x
horizontally (missing the right half of the field of view completely).

Therefore we must additionally set the horizontal binning mode. There
is only marginal change in output quality and noise levels.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Fixes: f075893e9b0e ("drivers: media: imx477: Disable the scaler")
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>