clk: microchip: mpfs: setup for using other mss pll outputs
commit dd3f761d08ce9066adadc2c2288bf6f1660a01c8 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Now that the MSSPLL is split, and the "postdiv" divider of the
cpu/AHB/AXI bus clock is represented by its own "hw" struct, make the
shifts, register offset and width a parameter of the initialisation
macro, rather than using defines that only work for one of the four
outputs.
Configuring this at initialisaion paves the way for using the other
three output clocks, where the register offset, and the bit shift
within that register, will differ.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>