defconfigs: Drop FB_UDL from all Pi defconfigs
commit 93e87e5c55e319be1658317c23b1cf4b62bba49d from
https://github.com/raspberrypi/linux.git rpi-6.6.y

CONFIG_FB_UBL is the fbdev driver for DisplayLink USB displays.
CONFIG_UDL is the DRM driver for DisplayLink USB displays.

Both attach themselves to the same USB device alias, so both
get loaded for any attached display. You now have 2 drivers for
the same hardware, and a race as to which one will load.

DRM emulates the framebuffer interface, so disable the fbdev
driver.

https://github.com/raspberrypi/bookworm-feedback/issues/133

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>