input: goodix: Include I2C details in names for the devices
commit a97af221c8c31e5ec21f46064bfd3f176248a14a from
https://github.com/raspberrypi/linux.git rpi-6.6.y

libinput uses the input device name alone. If you have two
identical input devices, then there is no way to differentiate
between them, and in the case of touchscreens that means no
way to associate them with the appropriate display device.

Add the I2C bus and address to the start of the input device
name so that the name is always unique within the system.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>