riscv: dts: microchip: add missing properties for fabric-bus 4
commit d483e58bf85e1645d4145d28b4416fa551d96648 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

ranges is a required property for simple-bus, so add an empty one as
there is no translation here.
The bus has a unit address, so it requires a reg property too. The only
thing actually on this bus is the athena crypto, so the reg property
here matches that of the crypto 1:1.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>