ARM: dts: Add CM5 DTS support
commit c2ec9eb29f257c1dbc40e9b5e1ba93e62524b3c2 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The CM5 is a platform that will appear in multiple boards, each of
which may have different connectivity. Split the CM5 DTS into a common
cm5.dtsi and board-specific dts files, where the CM5 DTS file (the one
loaded by the firmware by default) is an alias for the CM5IO DTS file.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>