media: ov64a40: Report the full crop rectangle
commit c3ef8e5cc3dc27a6f3d27794fa2fd0eeef0d162e from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The analogue crop rectangle defined in the sensor supported modes
reports the index of the last pixel in the horizontal/vertical
directions, instead of reporting the actual rectangle's width and
height.

Fix this by reporting the correct width and height and subtracting 1
pixel when applying the rectangle sizes to the sensor's registers.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>