Overlays:Add specific clk-producer/-consumer overlays for Hifiberry DAC+
As the easy switching of the I2S module bewteen clock producer/consumer
on the PI5 is not possible, two specific DT-overlays are introduced.
The DAC+PRO boards with onboard clocks use the -PRO overlay, the boards
without oscillators the -STD version.
The "hifiberry-dacplus,slave" parameter in the -STD overlay disables
the automatic clock detection inside the hifiberry-dacplus driver.

The former hifiberry-dacplus overlay is kept for compatibility but
will be deprecated.

Signed-off-by: j-schambacher <joerg@hifiberry.com>