dtoverlays: Fixup pendown gpio polarity for ads7846 users
commit c895cecf59190b45b8cfe3ad3f8edf3a7267c7f0 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The driver has been converted to use gpiod, which will normalise
polarity based on DT.

The piscreen overlay (and others) incorrectly defines the pendown
GPIO as being ACTIVE_HIGH (0), althought triggering on the high-low
edge for pen down. It therefore tries reading the pen position when
not being touched, and stops when it is touched.

Tested with piscreen and ads7846 overlays. Also fixed on others
where the interrupt says high->low but the polarity was ACTIVE_HIGH.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>