ARM: dts: pi400: Force stdout-path to serial0
commit b0ec72e3651aeffca2e31849e6f8d264f933270e from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The upstream Pi DTS files select serial1 for stdout-path, but because of
the way the downstream kernel numbers UARTS, serial0 is required in the
downstream DTS. Using the wrong value breaks U-boot.

In the 6.6 kernel the downstream Pi 400 DTS was made closer to upstream,
inheriting most of its content from the Pi 4B DTS, but because of the
order of inclusion it lost the override for stdout-path. Restore that
override.

See: https://github.com/raspberrypi/firmware/issues/1875
Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>