dtoverlays: Fix noints mode of mcp23017
commit 384ad7fd45ab61b8260f400a069d2ca5e7dffe82 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

noints mode disables 2 fragments that configure a GPIO to
be used for the interrupt line from the MCP23017, but fails
to remove the pinctrl-0 reference or pinctrl-names.
It therefore fails to load due to an invalid phandle.

Move the pinctrl-0 and pinctrl-names properties so they
also get disabled by the noints override.

https://forums.raspberrypi.com/viewtopic.php?t=370907

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>