regulator: rpi_panel_v2: Add delay on I2C reads
commit a649152e16d5ca0714d3959fa38ed1ae639ff179 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

As with many microcontrollers, the panel wants to use clock
stretching during a read so that the appropriate return value
can be generated and programmed into the hardware.
With Pi0-3 really not supporting clock stretching, this resulted
in the panel firmware dying.

Insert a delay between the write and the read to give the firmware
a chance to generate the relevant return value.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>