ARM: dts: Move rpi-otp nodes onto a dedicated bus
commit a2fb37fbc6cc3f92c3ea2e1d8f4e15d924ac860c from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The rpi-otp driver uses a virtualised, OTP-relative addressing scheme.
However, when instance nodes are children of "/soc" they appear to be
addressable directly by the host, which is wrong (but not in a way which
causes an error unless one goes looking for one).

Add a wrapper (bus) node without a "ranges" property to make the
separation clear.

See: https://github.com/raspberrypi/linux/issues/6196

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>