drivers: mmc: core: handle card-removal when running CQE recovery
commit 22b6295e3b0818583bf523d4b14ca7e0308d6861 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Recovery claims the MMC card so the card-detect work gets significantly
delayed - leading to lots of error recovery loops that can never do
anything but fail.

Explicitly detect the card after CQE has halted and bail if it's not
there.

Also ratelimit a not-very-descriptive warning - one occurrence in dmesg
is enough to signal that something is amiss.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>