Revert "riscv: dts: microchip: update ihc node in beaglev-fire dts"
commit 779d253615a8238fb85c2471154fb62a19af1bb8 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

This reverts commit 8fa998aa30a68749b9816407fda8db8c44369eba.

Do not use the IPC mailbox driver, since the MIV-IHC v2 Soft-IP core
won't be publically available in the Libero IP catalog by the time the
2024.06 PolarFire SoC Linux release happens. This revert will be undone
in a subsequent release once the IP is available in the Libero Catalog.
Continue using the IHC mailbox driver with the MIV-IHC subsystem
available in the icicle kit reference design in the meantime.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>