Revert "riscv: dts: microchip: update ihc and rproc nodes to use ipc driver"
commit 991136e33009f982601038cd3b4be4c53664918f from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

This reverts commit 2899777c97fe20bbdfde02931914ef45a01965c9.

Do not use the IPC mailbox driver, since the MIV-IHC v2 Soft-IP core
won't be publically available in the Libero IP catalog by the time the
2024.06 PolarFire SoC Linux release happens. This revert will be undone
in a subsequent release once the IP is available in the Libero Catalog.
Continue using the IHC mailbox driver with the MIV-IHC subsystem
available in the icicle kit reference design in the meantime.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>