riscv: dts: microchip: use miv-ihc mailbox IP in mpfs disco kit
commit c45c4d30d3e66197ae20e1a3a9bfb3b73ff3b300 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Do not use the IPC mailbox, since the MIV-IHC v2 Soft-IP core won't be
publically available in the Libero IP catalog by the time the 2024.06
PolarFire SoC Linux release happens. This revert will be undone in a
subsequent release once the IP is available in the Libero Catalog.
Continue using the IHC mailbox IP with the MIV-IHC subsystem
available in the icicle kit reference design in the meantime.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>