riscv: dts: microchip: reduce hss buffer size in the mpfs disco kit dts
commit d8ef60b3b380b4d53b97a4217d420665150bb617 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Reduce the hss buffer reserved memory region size to 2MiB, since it
was found that the hss buffer memory was not being reserved by the
kernel, causing the disco kit to not boot. This happened as memblock
reserves regions from the top of memory downwards and the first
allocation takes place without handling the reserved memory, which
reserves the top 4 MiB of memory.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>