spi: dw: don't immediately kill DMA transfers if an error occurs
commit bd2d5021ba5cd668d8fe2f8f4fe33d89cdb9235e from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Disabling the peripheral resets controller state which has a dangerous
side-effect of disabling the DMA handshake interface while it is active.
This can cause DMA channels to hang.

The error recovery pathway will wait for DMA to stop and reset the chip
anyway, so mask further FIFO interrupts and let the transfer finish
gracefully.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>