drm: panel: ili9881: Add option to reconfigure setup commands
commit 8f4e42a1f0e6be2448fd2d5ba1f6e1ee9802c4c4 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The driver is typically asking for LP commands, but then tries
to send set_display_[on|off] from enable/disable when the host
will be in HS mode.
It also sends shutdown commands just before it asserts reset and
disables the regulator, which is rather redundant.

Add an option to configure these two choices from the panel_desc.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>