dmaengine: dw-axi-dmac: Allow client-chosen width
commit fa478c29cc5c23ff083fe34d2309265fee9acaa9 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

For devices where transfer lengths are not known upfront, there is a
danger when the destination is wider than the source that partial words
can be lost at the end of a transfer. Ideally the controller would be
able to flush the residue, but it can't - it's not even possible to tell
that there is any.

Instead, allow the client driver to avoid the problem by setting a
smaller width.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>