drm/gem: Create shmem GEM object in a given mountpoint
commit e63bfca26dee77f0ddaa674ff9d4338c771d98c5 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Commit be431dfec976e553a08883e26d0d0cc2598a8dfa upstream

Create a function `drm_gem_shmem_create_with_mnt()`, similar to
`drm_gem_shmem_create()`, that has a mountpoint as a argument. This
function will create a shmem GEM object in a given tmpfs mountpoint.

This function will be useful for drivers that have a special mountpoint
with flags enabled.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240923141348.2422499-7-mcanal@igalia.com
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>