drm/v3d: Support Big/Super Pages when writing out PTEs
commit 709ff25f55abbcdd75ba89fce824f929dac3a761 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Commit e4c17720262f394f6320a2b6e17a128bfdabb37c upstream

The V3D MMU also supports 64KB and 1MB pages, called big and super pages,
respectively. In order to set a 64KB page or 1MB page in the MMU, we need
to make sure that page table entries for all 4KB pages within a big/super
page must be correctly configured.

In order to create a big/super page, we need a contiguous memory region.
That's why we use a separate mountpoint with THP enabled. In order to
place the page table entries in the MMU, we iterate over the 16 4KB pages
(for big pages) or 256 4KB pages (for super pages) and insert the PTE.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240923141348.2422499-9-mcanal@igalia.com
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>