drivers: media: bcm2835_isp: Cache LS table dmabuf
commit 615f539982dde45108372a08ab3474380724fb33 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Clients such as libcamera do not change the LS table dmabuf on every
frame. In such cases instead of mapping/remapping the same dmabuf on
every frame to send to the firmware, cache the dmabuf once and only
update and remap if the dmabuf has been changed by the userland client.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>