dts: bcm2712-ds: Restore gpio-direct to gio_aon
commit df22f77509885e0e3087d87fe31ec56dd1b7f59e from
https://github.com/raspberrypi/linux.git rpi-6.12.y

gpio-direct mode is a modification of the brcmstb GPIO driver that
makes it play nicely with the userspace pinctrl utility. The mode
forces the drive to read its state from the hardware each time, rather
than relying on cached state. Doing so slightly reduces performance,
but this is not a heavily used code path.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>