media: i2c: imx415: Link frequencies are not exclusive to num lanes
commit 5028631e8b8c27ec850066610eb221d1b53385ee from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The link frequencies are equally valid in 2 or 4 lane modes, but
they change the hmax_min value for the mode as the MIPI block
has to have sufficient time to send the pixel data for each line.

Remove the association with number of lanes, and add hmax_min
configuration for both lane options.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>