ASoC: allo-piano-dac-plus: Remove pointless code
commit 73fc55e94436ee3f86790c9a72ff3bd46c69d0fe from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The codec control Digital Playback Volume is one of the controls deleted
by the allo-piano-dac-plus driver. It is effectively replaced by the
soundcard controls Master Playback Volume and Subwoofer Playback Volume.

Delete the code that sets the volume limit on those codec controls - the
limits on the soundcard volume controls are sufficient.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>