mmc: bcm2835: Add downstream overclocking support
commit 6bd57cee7237ce1c5f2c55cdeeb63611b3789ed8 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The principal differences between the downstream SDHOST driver and the
version accepted upstream driver are that the upstream version loses the
overclock support and DMA configuration via DT, but gains some tidying
up (and maintenance by the upstream devs).

Add the missing features (with the exception of the low-overhead logging)
as a patch to the upstream driver.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>