dtoverlays: adds support for Hifiberry ADC8x to the DAC8x
commit 077576fbf1bc2340b5d5b1a4f19df4743f0c3d01 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Allows the usage of ADC8x stacked on top of the DAC8x.
Activates all I2S pins and uses now the dummy-dai instead
of the formerly used pcm5102 to allow the use of a
capture device, too. The simple card driver will
probe for the ADC8x and may activate the 8 channel
capture. Uses GPIO5 for detection.

Signed-off-by: j-schambacher <joerg@hifiberry.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>