drm/vc4: tests: Update pv muxing tests now DSI1 is more flexible
commit ec4fc7d3c618219c2fd1858cbe14b7b5b920acf0 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The tests on vc4 (BCM2835-7) were checking for DSI1 muxing being
to restricted channel 2, and therefore muxing with TXP was impossible.

As we no longer have that restriction, update the capabilities
defined for DSI1, move the tests that used to be impossible to the
valid list, and extend for additional combinations that are now
possible.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>