spi: rp2040-gpio-bridge: probe: Cfg fast_xfer clk
commit 672715d5c4d5da3b77ea8b0c31961052ee355c33 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

Fast transfer mode requires that the first bit of data is clocked with a
rising edge. This can cause extra bits of data to be clocked on hardware
where the clock signal uses a pull-up. This change ensures that clk is
driven low before fast data transfer mode is entered.

Signed-off-by: Richard Oliver <richard.oliver@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>