DT: bcm2711/bcm2712: use upstream property for controlling pcie clkreq
commit 57bcfa6da45cec1329d72a95f3620255036b596d from
https://github.com/raspberrypi/linux.git rpi-6.12.y

L1 sub-state support and clkreq control are intermingled in the hardware,
and upstream now have a tri-state property to control this. The default
behaviour is now to enable clkreq control if the property is absent.

Default Pi 5 PCIex1 to "safe" to avoid breaking expansion boards without
the CLKREQ# line connected. Make this explicit in the pciex1-compat-pi5
overlay, and remove the obsolete CM4 property.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>