overlay: Rework/fix the Pi 5 MIDI UART overlays
commit 90bd324fb23dd2f4e8d53a54e12bcd99073e91db from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The old Pi 5 MIDI UART overlays replaced the reference to the UART's
source clock with a fixed clock (essentially an integer dressed up like
a clock). This is fragile, in that the fixed clock must be updated if
the source clock's (fixed) frequency is ever changed, but it also relies
on the source clock having been enabled, which is likely not to be the
case unless another RP1 UART is being actively used.

Switching to a fixed-factor-clock solves both of those problems.

See: https://forums.raspberrypi.com/viewtopic.php?t=385906

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>